+++ CHANGES/6352.feature
@@ -1,0 +1,1 @@
+``X509CertGuard.ca_certificate`` is now stored in the database and not on the filesystem.
