+++ pulp_certguard/app/models.py
@@ -6,7 +6,6 @@ from django.db import models
 
 from OpenSSL import crypto as openssl
 
-from pulpcore.plugin import storage
 from pulpcore.plugin.models import ContentGuard
 
 from pulp_certguard.app.utils import get_rhsm
@@ -20,52 +19,10 @@ except ImportError:
 log = getLogger(__name__)
 
 
-class RHSMCertGuard(ContentGuard):
-    """
-    A content-guard validating on a RHSM Certificate validated by `python-rhsm`.
-
-    A Certificate Authority certificate to trust is required with each RHSMCertGuard created. With
-    each request, the client certificate is first checked if it is signed by this CA cert. If not,
-    it's untrusted and denied regardless of its paths.
-
-    After determining the client certificate is trusted, the requested path is checked against the
-    named paths in the certificate. A request is permitted if the current request path is a prefix
-    of a path declared in the trusted RHSM Client Certificate.
-
-    Fields:
-        rhsm_certificate (models.TextField): The RHSM Ccertificate used to validate the client
-            certificate at request time.
-    """
-
-    TYPE = 'rhsm'
+class BaseCertGuard(ContentGuard):
+    """A Base class all CertGuard implementations should derive from."""
 
     ca_certificate = models.TextField()
-
-    def __init__(self, *args, **kwargs):
-        """Initialize a RHSMCertGuard and ensure this system has python-rhsm on it."""
-        get_rhsm()  # Validate that rhsm is installed
-        super().__init__(*args, **kwargs)
-
-    class Meta:
-        default_related_name = "%(app_label)s_%(model_name)s"
-
-    def permit(self, request):
-        """
-        Validate the client cert is trusted and asserts a path that is prefix of the requested path.
-
-        Args:
-            request: The request from the user.
-
-        Raises:
-            PermissionError: If the request path is not a subpath of a path named in the
-                certificate, or if the client certificate is not trusted from the CA certificated
-                stored as `ca_certificate`.
-        """
-        get_rhsm()
-        unquoted_certificate = unquote(self._get_client_cert_header(request))
-        self._ensure_client_cert_is_trusted(unquoted_certificate)
-        rhsm_cert = self._create_rhsm_cert_from_pem(unquoted_certificate)
-        self._check_paths(rhsm_cert, request.path)
 
     @staticmethod
     def _get_client_cert_header(request):
@@ -106,6 +63,55 @@ class RHSMCertGuard(ContentGuard):
         except openssl.Error as exc:
             raise PermissionError(str(exc))
 
+    class Meta:
+        abstract = True
+
+
+class RHSMCertGuard(BaseCertGuard):
+    """
+    A content-guard validating on a RHSM Certificate validated by `python-rhsm`.
+
+    A Certificate Authority certificate to trust is required with each RHSMCertGuard created. With
+    each request, the client certificate is first checked if it is signed by this CA cert. If not,
+    it's untrusted and denied regardless of its paths.
+
+    After determining the client certificate is trusted, the requested path is checked against the
+    named paths in the certificate. A request is permitted if the current request path is a prefix
+    of a path declared in the trusted RHSM Client Certificate.
+
+    Fields:
+        rhsm_certificate (models.TextField): The RHSM Ccertificate used to validate the client
+            certificate at request time.
+    """
+
+    TYPE = 'rhsm'
+
+    def __init__(self, *args, **kwargs):
+        """Initialize a RHSMCertGuard and ensure this system has python-rhsm on it."""
+        get_rhsm()  # Validate that rhsm is installed
+        super().__init__(*args, **kwargs)
+
+    class Meta:
+        default_related_name = "%(app_label)s_%(model_name)s"
+
+    def permit(self, request):
+        """
+        Validate the client cert is trusted and asserts a path that is prefix of the requested path.
+
+        Args:
+            request: The request from the user.
+
+        Raises:
+            PermissionError: If the request path is not a subpath of a path named in the
+                certificate, or if the client certificate is not trusted from the CA certificated
+                stored as `ca_certificate`.
+        """
+        get_rhsm()
+        unquoted_certificate = unquote(self._get_client_cert_header(request))
+        self._ensure_client_cert_is_trusted(unquoted_certificate)
+        rhsm_cert = self._create_rhsm_cert_from_pem(unquoted_certificate)
+        self._check_paths(rhsm_cert, request.path)
+
     @staticmethod
     def _create_rhsm_cert_from_pem(unquoted_certificate):
         try:
@@ -122,7 +128,7 @@ class RHSMCertGuard(ContentGuard):
             raise PermissionError(msg)
 
 
-class X509CertGuard(ContentGuard):
+class X509CertGuard(BaseCertGuard):
     """
     A content-guard that authenticates the request based on a client provided X.509 Certificate.
 
@@ -133,137 +139,19 @@ class X509CertGuard(ContentGuard):
 
     TYPE = 'x509'
 
-    def _certpath(self, name):
-        return storage.get_tls_path(self, name)
-
-    ca_certificate = models.FileField(max_length=255, upload_to=_certpath)
-
     def permit(self, request):
-        """Authorize the specified web request.
+        """
+        Validate the client cert is trusted.
 
         Args:
-            request (aiohttp.web.Request): A request for a published file.
+            request: The request from the user.
 
         Raises:
-            PermissionError: When the request cannot be authorized.
-
+            PermissionError: If the client certificate is not trusted from the CA certificated
+                stored as `ca_certificate`.
         """
-        ca = self.ca_certificate.read()
-        validator = X509Validator(ca.decode('utf8'))
-        validator(request)
+        unquoted_certificate = unquote(self._get_client_cert_header(request))
+        self._ensure_client_cert_is_trusted(unquoted_certificate)
 
     class Meta:
         default_related_name = "%(app_label)s_%(model_name)s"
-
-
-class X509Validator:
-    """An X.509 certificate validator."""
-
-    CERT_HEADER_NAME = 'X-CLIENT-CERT'
-
-    @staticmethod
-    def format(pem):
-        """Ensure the PEM encoded certificate is properly formatted.
-
-        The certificate is passed as an HTTP header which does not permit newlines.
-
-        Args:
-            pem (str): A PEM encoded certificate.
-
-        Returns:
-            str: A properly PEM formatted certificate.
-
-        """
-        header = '-----BEGIN CERTIFICATE-----'
-        footer = '-----END CERTIFICATE-----'
-        body = pem.replace(header, '')
-        body = body.replace(footer, '')
-        body = body.strip(' \n\r')
-        return '\n'.join((header, body, footer))
-
-    @staticmethod
-    def load(pem):
-        """Load the PEM encoded certificate.
-
-        Encapsulates complexity of OpenSSL.
-
-        Args:
-            pem (str): A PEM encoded certificate.
-
-        Returns:
-            openssl.X509: The loaded certificate.
-
-        Raises:
-            ValueError: On load failed.
-
-        """
-        try:
-            return openssl.load_certificate(openssl.FILETYPE_PEM, buffer=X509Validator.format(pem))
-        except Exception as le:
-            raise ValueError(str(le))
-
-    def client_certificate(self, request):
-        """Extract and load the client certificate passed in the X-CLIENT-CERT header.
-
-        Args:
-            request (aiohttp.web.Request): A request for a published file.
-
-        Returns:
-            openssl.X509: The loaded certificate.
-
-        Raises:
-            KeyError: When the client certificate header has not
-                been passed in the request.
-
-        """
-        name = self.CERT_HEADER_NAME
-        try:
-            certificate = request.headers[name]
-        except KeyError:
-            reason = _('HTTP header "{h}" not found.').format(h=name)
-            raise KeyError(reason)
-        else:
-            return X509Validator.load(certificate)
-
-    def __init__(self, ca_certificate):
-        """Inits a new validator.
-
-        Args:
-            ca_certificate (str): A PEM encoded CA certificate.
-
-        """
-        self.ca_certificate = self.load(ca_certificate)
-
-    @property
-    def store(self):
-        """A X.509 certificate (trust) store.
-
-        Returns:
-            openssl.X509Store: A store containing the CA certificate.
-
-        """
-        store = openssl.X509Store()
-        store.add_cert(self.ca_certificate)
-        return store
-
-    def __call__(self, request):
-        """Validate the client X.509 certificate passed in the request.
-
-        Args:
-            request (aiohttp.web.Request): A request for a published file.
-
-        Raises:
-            PermissionError: When validation the client certificate
-                cannot be validated.
-
-        """
-        try:
-            context = openssl.X509StoreContext(
-                certificate=self.client_certificate(request),
-                store=self.store,
-            )
-            context.verify_certificate()
-        except openssl.X509StoreContextError:
-            raise PermissionError(_('Client certificate cannot be validated.'))
-        except Exception as pe:
-            raise PermissionError(str(pe))
