+++ pulp_certguard/tests/functional/api/test_x509_certguard.py
@@ -1,217 +1,55 @@
-import unittest
-from random import choice
-from requests import HTTPError
-from pulp_smash import api, config, utils
+import uuid
 
-from pulp_smash.pulp3.utils import (
-    download_content_unit,
-    gen_distribution,
-    gen_repo,
-    sync,
-)
+from pulpcore.client.pulp_certguard import CertguardX509CertGuard, ContentguardsX509Api
+
+from pulp_certguard.tests.functional.api.base import BaseCertGuard, CommonDenialTestsMixin
 from pulp_certguard.tests.functional.constants import (
-    X509_CERT_CA_FILE_PATH,
-    X509_CERT_CLIENT_FILE_PATH,
-    X509_CONTENT_GUARD_PATH,
+    X509_BASE_PATH,
+    X509_CA_CERT_FILE_PATH,
+    X509_CLIENT_CERT_FILE_PATH,
+    X509_UNTRUSTED_CLIENT_CERT_FILE_PATH,
 )
-from pulp_file.tests.functional.constants import (
-    FILE_DISTRIBUTION_PATH,
-    FILE_REMOTE_PATH,
-    FILE_REPO_PATH,
-)
-from pulp_file.tests.functional.utils import (
-    create_file_publication,
-    gen_file_remote,
-    get_file_content_paths
+from pulp_certguard.tests.functional.utils import (
+    gen_certguard_client,
+    set_distribution_base_path_and_download_a_content_unit_with_cert,
 )
 
 
-class X509CertGuardTestCase(unittest.TestCase):
+class X509CertGuardTestCase(BaseCertGuard, CommonDenialTestsMixin):
     """Api tests for X509CertGard."""
 
-    @classmethod
-    def setUpClass(cls):
-        """
-        Initialize common use entities and variables.
-
-        0. Prepare artifacts (ca and client certs)
-        1. Create self.contentguard for ca.pem located in CERT_CA_PATH.
-        2. create a repo.
-        3. create a file remote.
-        4. Sync it.
-        5. Create a Publication.
-
-        """
-        cls.cfg = cfg = config.get_config()
-        cls.client = client = api.Client(cfg, api.page_handler)
-        cls.teardown_cleanups = []
-
-        # The certificate header must have no \n chars
-        with open(X509_CERT_CLIENT_FILE_PATH, 'r') as cert_client_file:
-            cls.client_cert = str(cert_client_file.read()).replace('\n', '')
-
-        with utils.ensure_teardownclass(cls):
-            # 1. Create X.509 ContentGuard
-            with open(X509_CERT_CA_FILE_PATH, 'rb') as cert_ca_file:
-                cls.x509_certguard = client.post(
-                    X509_CONTENT_GUARD_PATH,
-                    data={'name': utils.uuid4()},
-                    files={'ca_certificate': cert_ca_file}
-                )
-                cls.teardown_cleanups.append(
-                    (client.delete, cls.x509_certguard['pulp_href'])
-                )
-
-            # 2. Create a repo
-            _repo = client.post(FILE_REPO_PATH, gen_repo())
-            cls.teardown_cleanups.append((cls.client.delete, _repo['pulp_href']))
-
-            # 3. Create a remote
-            cls.remote = client.post(FILE_REMOTE_PATH, gen_file_remote())
-            cls.teardown_cleanups.append(
-                (cls.client.delete, cls.remote['pulp_href'])
-            )
-
-            # 4. Sync and read synced repo
-            sync(cfg, cls.remote, _repo)
-            cls.repo = client.get(_repo['pulp_href'])
-
-            # 5. Create a publication
-            cls.publication = create_file_publication(cfg, cls.repo)
-            cls.teardown_cleanups.append(
-                (cls.client.delete, cls.publication['pulp_href'])
-            )
-
-    def test_negative_download_protected_content_without_keys(self):
-        """
-        Assert content protected by X.509 Cert Guard cannot be downloaded.
-
-        1. Create a protected distribution using the x.509 Cert Guard.
-        2. Assert content cannot be downloaded without cert and key.
-        """
-        # Create a protected distribution
-        distribution = self.client.using_handler(api.task_handler).post(
-            FILE_DISTRIBUTION_PATH,
-            gen_distribution(
-                publication=self.publication['pulp_href'],
-                content_guard=self.x509_certguard['pulp_href']
-            )
-        )
-        self.addCleanup(self.client.delete, distribution['pulp_href'])
-
-        # Pick a filename
-        unit_path = choice(get_file_content_paths(self.repo))
-
-        # Try to download it without the X-CLIENT-CERT header set
-        with self.assertRaises(HTTPError):
-            download_content_unit(self.cfg, distribution, unit_path)
-
-    def test_positive_download_protected_content_with_keys(self):
-        """
-        Assert content protected by X.509 Cert Guard can be downloaded.
-
-        1. Create a protected distribution using the self.contentguard.
-        2. Assert content can be downloaded using the proper cert and key.
-        """
-        # Create a protected distribution
-        distribution = self.client.using_handler(api.task_handler).post(
-            FILE_DISTRIBUTION_PATH,
-            gen_distribution(
-                publication=self.publication['pulp_href'],
-                content_guard=self.x509_certguard['pulp_href']
-            )
-        )
-        self.addCleanup(self.client.delete, distribution['pulp_href'])
-
-        # Pick a filename
-        unit_path = choice(get_file_content_paths(self.repo))
-
-        # Try to download it passing the proper X-CLIENT-CERT
-        download_content_unit(
-            self.cfg,
-            distribution,
-            unit_path,
-            headers={'X-CLIENT-CERT': self.client_cert}
-        )
-
-    def test_positive_add_x509_certguard_to_existing_distribution(self):
-        """
-        Assert adding X.509CertGuard to existing distribution works well.
-
-        1. Create a distribution without protection
-        2. Assert content can be downloaded
-        3. Add X.509CertGuard to the distribution
-        4. Assert content cannot be downloaded without key
-        5. Assert content can be downloaded with key
-        """
-        # 1 unprotected distribution
-        distribution = self.client.using_handler(api.task_handler).post(
-            FILE_DISTRIBUTION_PATH,
-            gen_distribution(publication=self.publication['pulp_href'])
-        )
-        self.addCleanup(self.client.delete, distribution['pulp_href'])
-
-        # Pick a filename
-        unit_path = choice(get_file_content_paths(self.repo))
-
-        # Download it without certificate
-        download_content_unit(self.cfg, distribution, unit_path)
-
-        # Update distribution adding the guard
-        distribution = self.client.using_handler(api.task_handler).patch(
-            distribution['pulp_href'],
-            {'content_guard': self.x509_certguard['pulp_href']}
-        )
-
-        # Cannot download without key
-        with self.assertRaises(HTTPError):
-            download_content_unit(self.cfg, distribution, unit_path)
-
-        # Try to download it passing the proper X-CLIENT-CERT
-        download_content_unit(
-            self.cfg,
-            distribution,
-            unit_path,
-            headers={'X-CLIENT-CERT': self.client_cert}
-        )
-
-    def test_positive_remove_contentguard(self):
-        """
-        Assert that content can be download without X.509CertGuard if it is removed.
-
-        1. Create a protected distribution
-        2. Assert content cannot be downloaded without keys
-        3. Remove the X.509CertGuard
-        4. Assert content can be downloaded without keys
-        """
-        # Create a protected distribution
-        distribution = self.client.using_handler(api.task_handler).post(
-            FILE_DISTRIBUTION_PATH,
-            gen_distribution(
-                publication=self.publication['pulp_href'],
-                content_guard=self.x509_certguard['pulp_href']
-            )
-        )
-        self.addCleanup(self.client.delete, distribution['pulp_href'])
-
-        # Pick a filename
-        unit_path = choice(get_file_content_paths(self.repo))
-
-        # Try to download it without the X-CLIENT-CERT
-        with self.assertRaises(HTTPError):
-            download_content_unit(self.cfg, distribution, unit_path)
-
-        # Update distribution removing the guard
-        distribution = self.client.using_handler(api.task_handler).patch(
-            distribution['pulp_href'],
-            {'content_guard': None}
-        )
-
-        # Now content can be downloaded
-        download_content_unit(self.cfg, distribution, unit_path)
+    DENIALS_BASE_PATH = X509_BASE_PATH
+    UNTRUSTED_CLIENT_CERT_PATH = X509_UNTRUSTED_CLIENT_CERT_FILE_PATH
 
     @classmethod
-    def tearDownClass(cls):
-        """Clean class-wide variable."""
-        for cleanup_function, args in reversed(cls.teardown_cleanups):
-            cleanup_function(args)
+    def _setup_content_guard(cls):
+        # Create an X.509 Content Guard
+        certguard_client = gen_certguard_client()
+        cls.x509_content_guards_api = ContentguardsX509Api(certguard_client)
+
+        with open(X509_CA_CERT_FILE_PATH, 'r') as x509_ca_cert_data_file:
+            x509_ca_cert_data = x509_ca_cert_data_file.read()
+
+        x509_cert_guard = CertguardX509CertGuard(
+            name=str(uuid.uuid4()),
+            ca_certificate=x509_ca_cert_data
+        )
+        cls.x509_content_guard_data = cls.x509_content_guards_api.create(x509_cert_guard)
+        cls.teardown_cleanups.append(
+            (cls.x509_content_guards_api.delete, cls.x509_content_guard_data.pulp_href)
+        )
+        return cls.x509_content_guard_data.pulp_href
+
+    def test_allow_request_when_cert_is_trusted(self):
+        """
+        Assert a correctly configured client can fetch content.
+
+        1. Configure the distribution with an X.509 CertGuard.
+        2. Attempt to download content.
+        """
+        set_distribution_base_path_and_download_a_content_unit_with_cert(
+            self.distribution.pulp_href,
+            self.DENIALS_BASE_PATH,
+            self.repo.pulp_href,
+            X509_CLIENT_CERT_FILE_PATH
+        )
