+++ pulp_certguard/tests/functional/artifacts/x509/certificates/untrusted_client.pem
@@ -1,0 +1,18 @@
+-----BEGIN CERTIFICATE-----
+MIIC8jCCAdoCAgOtMA0GCSqGSIb3DQEBBQUAMDYxNDAyBgNVBAMMK3B1bHAzLXNv
+dXJjZS1mZWRvcmEzMS5sb2NhbGhvc3QuZXhhbXBsZS5jb20wHhcNMjAwNDAzMTky
+OTI5WhcNMzAwNDAxMTkyOTI5WjBHMTQwMgYDVQQDDCtwdWxwMy1zb3VyY2UtZmVk
+b3JhMzEubG9jYWxob3N0LmV4YW1wbGUuY29tMQ8wDQYDVQQKDAZjbGllbnQwggEi
+MA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQC+/beGWsHQJYbFV86jmi/FJOCe
++7+myDYWEZFFMu7IPvkoSQPC9OJh/U6IOhmXWqoIMENh118FzLYB9bAmZt/gbMtc
+4i6XKzHfgiHEGfM/s+CdqC0qNrxelydwt3ceBOh71chbalsALnAthwSHMn5VWOz1
+ym89LSuf0jlqchOqw6Vf/8t2PIhRagM13B0pcFeqCNv/jv6jwzgN7WuI1hhC9EAT
+KLm05Y0JDPXBb/Gq97X66PhDOoXso0i5yjHj7/KsFY1W4C8vY68l3epybJ7LJ7Kf
++tV2YX27LLx8gagdadUDopUqGV3nFDjoY6+MtVRp4JU9h3/8BZJzUQdihM7LAgMB
+AAEwDQYJKoZIhvcNAQEFBQADggEBAL9tT6vbbuycVyR3eBdzlZwMYMV++44EjYIp
+daRXTQRzrGJ6A9hNLkE38k1gM+vTL62UOFXA+xALuXXfxx7BswGg9Au0bCyODa8K
+LhuHAILVcFaRpW5HxQ56+oSGdt4tyKGwuZeBmfx0fWzP1RPML2WCFfHT4yPXz/+s
+c00/8fJhWt+W6lmc95l2UC2+PSrZSAgqCMpjdg3gTzP3v9d7N55UxTJ5TEtTK5X1
+HKrez9JYYkm8ec72xboOKljdfRF7s1lcIMQ7PLFrS7fQ1xVZTFmYdQnAITypPztM
+67Lxjxi0iGVQD3I+KxqzOj3AvgmO09H/QdNIbxJLiI/XyWTUFuc=
+-----END CERTIFICATE-----
